Updating CMakeLists for 64-bit VS2019 builds (for now, just targeting windows OS), using the Clang compiler. 1. WIN32 symbol is now defined, when the C/C++ compiler is invoked, only in 32-bit windows     builds (WIN64 is set instead in 64-bit builds) 2.  CLANG_UNDER_VS is set (and a symbol with the same name is defined when the C/C++ compiler is invoked)      (when CMake is being invoked from Visual Studio). When it is set     (a) _UWIN is defined and /GX is passed (when you're calling the VS-friendly wrapper for Clang,          /GX is equivalent to passing -fexceptions to the usual version of clang).     (b) /MP (which requests multi-processor compilation) is *not* passed     (c) /O2 is passed (in release builds) (Seems odd.  You'd have thought, /O3, surely?)     (d) In 64-bit builds, -m64 is supplied as a parameter to the compiler     (e) If MPI package isn't found (it doesn't seem to be) MPI_DIR is as passed to the compiler          (you'll need to download MS-MPI and point MPI_DIR at it) 3. Detection of 32-bit versus 64-bit binary is done sooner than it used to be (to facilitate 2(d) above).
Adjusting includes that cause problems for Windows/VS builds (stuff listed here indicates how VS is different):
1. Disabling #include <getopt.h> (e.g. in booster/booster.c)
2. Disabling #include <libgen.h> (e.g. in booster/version.h)
3. Enabling #include <winsock2.h> and <windows.h> (e.g. in main/main.cpp)
4. Adding a header file (systypes.h) in pll so as to disable #includes of <unistd.h> , <sys/times.h>,
    <sys/types.h>, <sys/time.h>
5. Including MPI headers when _IQTREE_MPI is defined, but _FINE_GRAIN_MPI is not.  The problem is that
    the pll files sort of assume MPI is set *or* pthreads is available.  But in Windows builds I don't want to add
    a dependency on a pthreads implementation, because there isn't a "standard" Microsoft one (and there
    *is* a "standard" Microsoft implementation of MPI).    As yet I've only built with MS-MPI version 10.1.2.
6. Not including "pthread.h" (or <pthread.h>) in CLANG_UNDER_VS builds (e.g. in pll/genericParallelization.h)
    (for the same reasons; I don't *yet* want a dependency on a non-MS Windows pthreads implementation ).

Adjusting conditional compilation directives so they recognize WIN64 as well as WIN32 (lots of files).
1. Warnings 4786 and 4291 aren't disabled, and vsnprintf isn't treated as a synonym for _vsnprintf
     (in ncl/ncl.h).
2. _mm_popcnt_u64 is mapped to _popcnt64 (there may be some backwards compatibility issues here
     in older versions of Visual Studio, also perhaps on 32-bit VS builds). (e.g. in pll/fastDNAparsimony.c)
     (and also in tree/phylokernel.h)
3. Faked S_ISDIR and S_ISREG (which aren't always available in WIndows builds) (e.g. in pll/hardware.c)
     (and also in utils/tools.cpp).
4. Did an "#undef max" in terraphast/include/terraces/definitions.hpp.

Silencing warnings from Visual Studio Builds:
1. Ensuring that members are initialized in constructors (e.g. SymTestResult in alignment.h)
2. C++ style casts (e.g. static_cast) where that was needed to silence conversion warnings.
3. Promoted the return type of MSetsBlock::getNSets() to size_t. (in nclextra/msetsblock.h).
4. Promoted the return type of SplitGraph::getNSplits() to size_t (in pda/splitgraph.h)
5. Promoted the retyrn type of PML::StateSpace::getNAllStates() to size_t (in phlo-yaml/statespace.h)
6. Initializing local variables that weren't set in all cases of switch statements (this happened a lot).
7. Renaming local variables called this (e.g. in fastDNAparsimony.c) to: here (keyword issue!).
8. Ensuring that string literals aren't assigned to (non-const) char*, but rather to const char* (e.g.
    in pll/genericParallelization.c).
9.. Declared an inline rax_malloc_string_copy (in pll/mem_alloc.h) to silence warnings about the use
     of strcpy (and simplify calling code that is repeatedly doing the same stuff to "mimic" the
     behaviour of the classic strdup function).
10. Removed unused variables (e.g. oldFreqs variable in autoProtein in pll/optimizeModel.c).
      Actually, that wasn't just unused, it was leaked!  So I killed it.
11. Renamed parameters that shadowed class members (e.g. terraces::constraint constructor in
     terraphast/include/terraces/constraints.hpp).
12. Changed teraphast/CMakeLists.txt compilation flags to exclude -Za when Clang is being used under
      Visual Studio (when CLANG_UNDER_VS is set).
13. Constructors in "diamond-inheritance" classes that didn't call all their superclass constructors now do.
      (I'm surprised this was being let go!) in utils/gzstream.h.
14. Removed the register keyword in matinv in whtest/eigen.c (register is... deprecated at least!  Probably
      even disallowed by now, in the latest C++ and C# standards).
15. Replaced K&R 1 function definitions (in zlib-1.2.7/inflate.c).  K&R 1 is so 1970s!  This no longer even
      compiles under Visual Studio.  Those sort of definitions went out of fashion in about 1980!

Patching a bug in NJMatrix::calculateScaledRowTotals (in utils/bionj2.cpp); it was calling reserve() rather than resize() on
a vector that it meant to... resize.

Note: utils/tools.cpp's getFIlesInDir is broken in windows builds (no "innards").
Which means that SuperAlignment::readPartitionDir will not work on Windows (it won't have
a list of the files in a directory).

+++ phylo-yaml/statespace.h
@@ -94,7 +94,7 @@ public:
     inline int getNStates() { return num_states; }
 
     /** get all number of states incl. missing/gap/ambiguous states */
-    inline int getNAllStates() { return states.size(); }
+    inline size_t getNAllStates() { return states.size(); }
 
     /**
      initialise from a state definition string
